refactor(auction): 경매 정산 로직 구조 개선 및 트랜잭션 안정성 확보
